fix(x11): fix rendering when DPI scaling is less than 1

+++ src/Uno.UI.Runtime.Skia.X11/X11SoftwareRenderer.cs
@@ -74,7 +74,7 @@ namespace Uno.WinUI.Runtime.Skia.X11
 				var scale = host.RootElement?.XamlRoot is { } root
 					? root.RasterizationScale
 					: 1;
-				canvas.Scale((int)scale);
+				canvas.Scale((float)scale);
 
 				if (host.RootElement?.Visual is { } rootVisual)
 				{
